Review tab - Move getFlavorValue function to utils
