+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/tabs/review-tab/general-tab.tsx
@@ -9,12 +9,7 @@ import { FormFieldReviewMemoRow } from '../../form/form-field-review-row';
 import { FormFieldType } from '../../form/form-field';
 import { iGetCommonData, iGetLoadedCommonData } from '../../selectors/immutable/selectors';
 import { getFieldValue } from '../../selectors/vm-settings';
-import { iGetRelevantTemplate } from '../../../../selectors/immutable/template/combined';
-import { toShallowJS } from '../../../../utils/immutable';
-import { VMTemplateWrapper } from '../../../../k8s/wrapper/vm/vm-template-wrapper';
-import { CUSTOM_FLAVOR } from '../../../../constants/vm';
-import { iGetFieldValue } from '../../selectors/immutable/field';
-import { getField } from './utils';
+import { getField, getFlavorValue } from './utils';
 import { VMSettings } from '../../redux/initial-state/types';
 import { getOS } from '../../selectors/combined';
 
@@ -23,29 +18,16 @@ import './review-tab.scss';
 const GeneralReviewConnected: React.FC<GeneralReviewConnectedProps> = (props) => {
   const { iVMSettings, iUserTemplates, iCommonTemplates, openshiftFlag, relevantOptions } = props;
 
-  const getFlavorValue = () => {
-    const flavor = iGetFieldValue(getField(VMSettingsField[VMSettingsField.FLAVOR], iVMSettings));
-    let cpuCores, memory;
-
-    if (flavor === CUSTOM_FLAVOR) {
-      cpuCores = iGetFieldValue(getField(VMSettingsField.CPU, iVMSettings));
-      memory = iGetFieldValue(getField(VMSettingsField.MEMORY, iVMSettings));
-    } else {
-      const template_ = toShallowJS(
-        iGetRelevantTemplate(iUserTemplates, iCommonTemplates, relevantOptions),
-      );
-      const template = new VMTemplateWrapper(template_, true).init().clearRuntimeMetadata();
-
-      cpuCores = template.getCPU()?.cores;
-      memory = template.getMemory();
-    }
-
-    return `${flavor}: ${cpuCores} CPU, ${memory}`;
-  };
+  const flavorValue = getFlavorValue({
+    iVMSettings,
+    iUserTemplates,
+    iCommonTemplates,
+    relevantOptions,
+  });
 
   const osName =
     getOS({
-      osID: getFieldValue(iVMSettings.toJS() as VMSettings, VMSettingsField.OPERATING_SYSTEM),
+      osID: getFieldValue(iVMSettings as VMSettings, VMSettingsField.OPERATING_SYSTEM),
       iUserTemplates,
       openshiftFlag,
       iCommonTemplates,
@@ -68,7 +50,7 @@ const GeneralReviewConnected: React.FC<GeneralReviewConnectedProps> = (props) =>
 
       <FormFieldReviewMemoRow
         field={getField(VMSettingsField.FLAVOR, iVMSettings)}
-        value={getFlavorValue()}
+        value={flavorValue}
       />
 
       <FormFieldReviewMemoRow
